UWP versions of D3D12 samples
All existing samples re-added as UWP projects.  To see the differences
between UWP versions and UWP versions, you can diff the Desktop and UWP
folders.  These are added as copies so that the newcomer can focus on
D3D12 code and not be distracted by platform code.  D3D12Multithreading
is not being ported to UWP because it consumes a 30+MB asset that we did
not want to duplicate.