Fix root signature issue in D3D12nBodyGravity sample
The sample has a compute shader that reads outside the bounds of the SRV
input. When DATA_STATIC flags are set, the driver may choose to optimize
the descriptor into a root descriptor which has no boundary information.
This means that reads outside the bounds of the SRV are undefined
instead of returning 0's. Changing the root signature flags to
DESCRIPTORS_VOLATILE to block this potential optimization.