Adding support for Shader Records with Descriptor Tables and Empty Top Level Acceleration Structures (#297)
This adds support for local root signatures/shader records with descriptor table handles as the parameter. This removes the need for the app having to do it's own bindless indexing to associate a ClosestHit shader with a texture.

Also adding tests for verifying the use of descriptor tables. These tests have a pass-through raygen trace against a null acceleration structure and invoke the varying miss shaders that test different varieties of descriptor tables pulled from the shader record.

Finally, modifying the MiniEngine sample to use descriptor tables as local root signature arguments, which cleans up the shader code nicely.