Fix DxilPatchingUnitTests in develop-dxr and develop-dxr-rs5 (#424)
The tests were failing due to multiple problems.

On develop-dxr, BuildEmptyTopLevelAccelerationStructure crashed during creation of an empty AS due to underflow when calculating some sizes

On develop-dxr-rs5, the unit test state object creation code was not correct and had records missing.
In another place, the unit test state object creation code was failing due to a real fallback layer bug. This change makes the test work around the bug because it is not crucial to what the test is validating. For tracking, I have opened #17895658.